Testing Updates (WORKING?!?!?)
Changed the detector to a square.
Converted the radius to 200 instead of 10.
Changed theta and phi to 32 arcminutes.
Changed the source to have x_coord for both x and y, so that it is offset at a diagonal.
Changed to generating 50k rays.
Stopped plotting all rays that were created.
Made the individual dots on the detector smaller, so that more can be visible at one time.